requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
flake8==7.0.0
Flask==3.0.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
peewee==3.17.1
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==3.0.1
wtf-peewee==3.0.5
WTForms==3.1.2
